+++ app/models/utility/scenario/units/definitions.rb
@@ -439,6 +439,7 @@ module Utility
               "semovente_da_75_18": { sz: 3, mv: 5, wpn: :"c_da_75_18", arm: [3, 2, 2] },
               "semovente_da_90_53": { sz: 4, mv: 4, wpn: :"cannone_da_90_53", arm: [3, 2, 2] },
               "autoblindo_41": { sz: 3, mv: 4, wpn: :"breda_m35", whl: true, trn: 1, arm: [1, 0, 0], trt: [1, 1, 1] },
+              "lancia_1zm": { sz: 3, mv: 3, wpn: :"breda_m37", whl: true, trn: 1, arm: [0, 0, 0], trt: [0, 0, 0] },
               # Japanese Armored Vehicles
               "type_94": { sz: 3, mv: 4, wpn: :"type_3_hmg", arm: [1, 0, 0], trt: [1, 0, 0] },
               "type_97_te_ke": { sz: 3, mv: 4, wpn: :"37mm_type_94", arm: [1, 0, 0], trt: [1, 1, 1] },